RNN VAD: pitch gain comparison, minor optimization
This optimization avoids that the expression
"xy / std::sqrt(1.f + xx * yy)" is evaluated inside a loop.
By storing a pitch gain as a GainRatio object, the pitch gain
comparison can be rewritten using 3 multiplications.

Bug: webrtc:10480

+++ modules/audio_processing/agc2/rnn_vad/features_extraction.cc
@@ -14,6 +14,7 @@
 
 #include "modules/audio_processing/agc2/rnn_vad/lp_residual.h"
 #include "rtc_base/checks.h"
+#include "rtc_base/numerics/safe_conversions.h"
 
 namespace webrtc {
 namespace rnn_vad {
@@ -69,7 +70,7 @@ bool FeaturesExtractor::CheckSilenceComputeFeatures(
   // into the output vector (normalization based on training data stats).
   pitch_info_48kHz_ = pitch_estimator_.Estimate(lp_residual_view_);
   feature_vector[kFeatureVectorSize - 2] =
-      0.01f * (static_cast<int>(pitch_info_48kHz_.period) - 300);
+      0.01f * (pitch_info_48kHz_.period - 300);
   // Extract lagged frames (according to the estimated pitch period).
   RTC_DCHECK_LE(pitch_info_48kHz_.period / 2, kMaxPitch24kHz);
   auto lagged_frame = pitch_buf_24kHz_view_.subview(
